+++ lib/home/home_screen.dart
@@ -2,25 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:gunpla_database/authscreen/login_screen.dart';
 import 'package:gunpla_database/backend/auth_provider.dart';
-
-class LeadingImage extends StatelessWidget {
-  final String url;
-
-  const LeadingImage(this.url);
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      decoration: BoxDecoration(
-        borderRadius: BorderRadius.circular(8.0),
-      ),
-      child: AspectRatio(
-        aspectRatio: 3 / 2,
-        child: Image.network(url),
-      ),
-    );
-  }
-}
+import 'package:gunpla_database/home/gunpla_list_page.dart';
 
 class HomeScreen extends StatelessWidget {
   final CollectionReference collectionReference =
@@ -29,48 +11,18 @@ class HomeScreen extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      appBar: AppBar(
-          leading: IconButton(
-            onPressed: () {
-              AuthClass().signOut();
-              Navigator.pushAndRemoveUntil(
-                  context,
-                  MaterialPageRoute(builder: (context) => LoginScreen()),
-                  (route) => false);
-            },
-            icon: const Icon(Icons.logout),
-          ),
-          title: const Text('Gunpla Database')),
-      body: SingleChildScrollView(
-          child: StreamBuilder(
-        stream: collectionReference.snapshots(),
-        builder: (BuildContext context, AsyncSnapshot<QuerySnapshot> snapshot) {
-          if (snapshot.hasData) {
-            var docs = snapshot.data.docs;
-            return ListView.builder(
-                primary: false,
-                shrinkWrap: true,
-                itemCount: docs.length,
-                itemBuilder: (_, index) {
-                  var document = docs[index];
-                  return ListTile(
-                    isThreeLine: true,
-                    leading: LeadingImage(document['image']),
-                    title: Text(document['name']),
-                    subtitle: Text(
-                      document['series'],
-                      maxLines: 2,
-                      overflow: TextOverflow.ellipsis,
-                    ),
-                    trailing: const Icon(Icons.chevron_right_sharp),
-                  );
-                });
-          }
-          return Center(
-            child: CircularProgressIndicator(),
-          );
-        },
-      )),
-    );
+        appBar: AppBar(
+            leading: IconButton(
+              onPressed: () {
+                AuthClass().signOut();
+                Navigator.pushAndRemoveUntil(
+                    context,
+                    MaterialPageRoute(builder: (context) => LoginScreen()),
+                    (route) => false);
+              },
+              icon: const Icon(Icons.logout),
+            ),
+            title: const Text('Gunpla Database')),
+        body: GunplaListTile());
   }
 }
